requirements.txt
logzero>=1.5.0
requests>=2.21
pyyaml~=5.4
importlib-metadata>=1.2.0; python_version < '3.8'
contextvars;python_version<"3.7"